Create README.md
I decided to put together a list of Superhero/Comic movies, just to see what the data tells me. Originally, I was going to introduce the variable of Comic sales within X years before the release of a title and Y years after the release of a title. After looking at the budget of some of these movies V the total worldwide income, you notice some very interesting things. It raises questions such as:

Which studio has made/lost the most money distributing titles?
What is the biggest percentage increase in profit when comparing the budget and the earnings?

These two questions sound similar, but they are significantly different. For example, you can look at the budget of a movie like Joker or Deadpool. These movies had a significantly low budget, but the worldwide box office saw a 1400+% increase in ticket sales when compared to the budget. You take a renowned movie, such as Avengers: Endgame, but when looking at the percentage increase it is not as economically successful in terms of money spent V money earned. I believe the increase was something like 600+% approx. Then you look at money lost by movies. This is where you look towards titles like Steel featuring Shaq, Catwoman with Halle Berry or the 1980s Supergirl movie. These movies appear to have lost tens of millions of dollars. Then this reveals some other factors that require observation. 

Are superhero movies in the past, where the main protagonist is of a different ethnicity or gender, more or less successful than they are now?

This is a reasonable question when you consider the uproar of gender or ethnicity swapped characters in the media, which has seen frequent publication in the media. The database is currently incomplete, but this is something that I am actively working on.